requirements.txt
asgiref==3.8.1
Django==5.0.6
django-filter==24.2
django-mail-templated==2.6.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
Markdown==3.6
pillow==10.3.0
PyJWT==2.8.0
setuptools==69.1.0
sqlparse==0.5.0
wheel==0.42.0
